Fix problem with error and addition file test

+++ TeshCompiler/res/binfile.java
@@ -1,6 +1,6 @@
 public class binfile {
     public String name;
-    public int error;
+    public long error;
     public String directory;
 
     public binfile() {
